# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/songs": {
      "get": {
        "tags": [
          "songs"
        ],
        "operationId": "list_songs",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number, 1-indexed. Defaults to 1.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "Items per page. Defaults to 20. The server enforces a maximum.",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paged list of songs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse_SongSummary"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "songs"
        ],
        "operationId": "create_song",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSongRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}": {
      "get": {
        "tags": [
          "songs"
        ],
        "operationId": "get_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "songs"
        ],
        "operationId": "update_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSongRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "songs"
        ],
        "operationId": "delete_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/songs/{id}/lyrics": {
      "get": {
        "tags": [
          "songs"
        ],
        "operationId": "get_song_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song lyrics",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LyricsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "songs"
        ],
        "operationId": "put_song_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLyricsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Lyrics saved"
          },
          "404": {
            "description": "Song not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "songs"
        ],
        "operationId": "delete_song_lyrics",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Lyrics removed"
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistInfo": {
        "type": "object",
        "description": "A minimal artist record embedded in song and performance responses.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "CreateSongRequest": {
        "type": "object",
        "description": "Request body for `POST /api/songs`.",
        "required": [
          "title",
          "artist_ids",
          "tags",
          "image_ids"
        ],
        "properties": {
          "artist_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "lyrics": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional inline lyrics content. Creates a lyrics row in a single round trip."
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongTagAssignment"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ImageInfo": {
        "type": "object",
        "description": "An image record embedded in song responses.",
        "required": [
          "id",
          "public_url"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "public_url": {
            "type": "string"
          }
        }
      },
...
      "LyricsResponse": {
        "type": "object",
        "description": "Response body for lyrics GET endpoints.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "PagedResponse_SongSummary": {
        "type": "object",
        "description": "Generic envelope returned by all paginated list endpoints.",
        "required": [
          "items",
          "total",
          "page",
          "per_page"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "description": "Lean song representation returned by list endpoints.\n\nContains enough to render a song card without a follow up request.",
              "required": [
                "id",
                "title",
                "date_added",
                "artists"
              ],
              "properties": {
                "artists": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ArtistInfo"
                  }
                },
                "date_added": {
                  "type": "string",
                  "format": "date-time"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "title": {
                  "type": "string"
                }
              }
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "The page that was returned, 1-indexed.",
            "minimum": 0
          },
          "per_page": {
            "type": "integer",
            "format": "int32",
            "description": "Items per page as applied by the server (may be less than requested if capped).",
            "minimum": 0
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching items across all pages.",
            "minimum": 0
          }
        }
      },
...
      "SongResponse": {
        "type": "object",
        "description": "Full song metadata returned by detail endpoints.\n\nExcludes lyrics, fetch those via `GET /api/songs/{id}/lyrics` on demand.",
        "required": [
          "id",
          "title",
          "date_added",
          "artists",
          "tags",
          "images"
        ],
        "properties": {
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistInfo"
            }
          },
          "date_added": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImageInfo"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagInfo"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "SongTagAssignment": {
        "type": "object",
        "description": "A tag paired with its kind for application to a song.",
        "required": [
          "tag_id",
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SongTagKind"
          },
          "tag_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SongTagKind": {
        "type": "string",
        "description": "Valid kind values when applying a tag to a song.",
        "enum": [
          "genre",
          "source",
          "language",
          "misc"
        ]
      },
      "TagInfo": {
        "type": "object",
        "description": "A tag embedded in song responses.",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "description": "Freeform category string (e.g. `\"genre\"`, `\"mood\"`)."
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateLyricsRequest": {
        "type": "object",
        "description": "Request body for lyrics PUT endpoints.",
        "required": [
          "content"
        ],
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
...
      "UpdateSongRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/songs/{id}`.\n\nLyrics are excluded, use `PUT /api/songs/{id}/lyrics` instead.",
        "required": [
          "title",
          "artist_ids",
          "tags",
          "image_ids"
        ],
        "properties": {
          "artist_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "image_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongTagAssignment"
            }
          },
          "title": {
            "type": "string"
          }
        }
      }
    }
  }
}